refactor: merge lsp & lsp2
