+++ config/nvim/init.lua
@@ -2,4 +2,4 @@ require("pkgs")
 require("built_in_settings")
 require("pkg_settings")
 require("maps")
-require("lsp")
+require("lsp").setup_all_lang_servers()
