--- config/nvim/lua/lsp.lua
@@ -1,1 +1,0 @@
-require("lsp2").setup_all_lang_servers()
